refactor: sperate the all section of edit profile page and page section  components  folder
