+++ section-components/index.js
@@ -1,0 +1,16 @@
+import BasicInformationSection from './edit-profile/BasicInformationSection';
+import EducationSection from './edit-profile/EducationSection';
+import WorkExperienceSection from './edit-profile/WorkExperienceSection';
+import SkillsSection from './edit-profile/SkillsSection';
+import SocialLinksSection from './edit-profile/SocialLinksSection';
+import ResumeSection from './edit-profile/ResumeSection';
+import LeftEditProfilePart from './edit-profile/LeftEditProfilePart';
+
+export {
+  BasicInformationSection,
+  EducationSection,
+  WorkExperienceSection,
+  SkillsSection,
+  ResumeSection,
+  LeftEditProfilePart,
+};
