+++ tsconfig.json
@@ -18,7 +18,7 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "section-components/index.js"],
   "exclude": ["node_modules"]
 }
